requirements.txt
dnspython>=2.3.0
click>=8.1.0
rich>=13.0.0
requests>=2.31.0 